requirements.txt
psutil
python-json-logger
redis
kafka-python